requirements.txt
asgiref==3.6.0
Django==4.2
django-filter==23.1
django-location-field==2.7.0
django-phonenumber-field==7.0.2
django-seed==0.3.1
djangorestframework==3.14.0
Faker==18.3.4
phonenumberslite==8.13.8
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
toposort==1.10
tzdata==2023.3